pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pedadog"
version = "0.1.0"
description = "Legal Pedagogy Analysis Package"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Legal Pedagogy Team"}
]
keywords = ["legal", "pedagogy", "analysis", "case-beliefs", "character-beliefs"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Researchers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "numpy>=1.23.5",
    "pandas>=2.2.0,<2.3",
    "scipy>=1.12.0,<1.13",
    "matplotlib>=3.8.2,<3.9",
    "seaborn>=0.13.2,<0.14",
    "scikit-learn>=1.6.1,<2",
    "typer>=0.9.0,<0.10",
    "rich>=13.7.0,<13.8",
    "python-dotenv",
    "anthropic>=0.49.0,<0.50",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "pre-commit>=4.2.0,<5",
]
ocr = [
    "marker-pdf>=1.7.2,<2",
]
jailbreak = [
    "jailbreakbench>=1.0.0,<2",
    "litellm>=1.20.0,<1.30.0",
    "tenacity>=8.2.3,<9.0.0",
]

[project.urls]
Homepage = "https://wherewith.ai"
Repository = "https://github.com/wherewithal/legal-pedagogy"

[tool.setuptools.packages.find]
where = ["."]
include = ["pedadog*"]

[tool.setuptools.package-data]
pedadog = ["*.txt", "*.md"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "TD", "FIX", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["E501", "COM812", "ISC001"]

[tool.pytest.ini_options]
testpaths = ["experiments", "pedadog"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=pedadog --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true